Convert Trough into a dumb view
Makes it more reusable. Now its parent component is responsible
for giving it state like isLoading and the Twats it'll display.

Also take care of mapping Redux state to props in the Dashboard
component, since Trough will be used for other things and
shouldn't be dealing with state that's only relevant to Dashboard.